backend/requirements.txt
aiohttp==3.6.2
aniso8601==7.0.0
appdirs==1.4.3
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
black==19.10b0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
click==7.1.1
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==2.2.11
django-channels==0.7.0
django-channels-graphql-ws==0.4.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-graphql-playground==1.0
entrypoints==0.3
flake8==3.7.9
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
mccabe==0.6.1
msgpack==0.6.2
multidict==4.7.5
oauthlib==3.1.0
pathspec==0.7.0
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2019.3
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
yarl==1.4.2
zope.interface==4.7.2
